chore(redis): remove redis from form web app session and replace with mongodb
This is because the Azure Redis instance cannot be put behind a VNet (without great expense) and
the private endpoint work doesn't seem to like Kubernetes. We are already using MongoDB in the
architecture so this is largely a simplification.
