+++ forms-web-app/package.json
@@ -24,7 +24,7 @@
     "@pins/common": "file:../common",
     "accessible-autocomplete": "^2.0.3",
     "compression": "^1.7.4",
-    "connect-redis": "^5.0.0",
+    "connect-mongodb-session": "^2.4.1",
     "cookie-parser": "~1.4.4",
     "date-fns": "^2.16.1",
     "express": "~4.16.1",
@@ -43,7 +43,6 @@
     "nunjucks-date-filter": "^0.1.1",
     "pino": "^6.7.0",
     "pino-http": "^5.3.0",
-    "redis": "^3.0.2",
     "uuid": "^8.3.1"
   },
   "devDependencies": {
