+++ forms-web-app/src/app.js
@@ -6,11 +6,10 @@ const cookieParser = require('cookie-parser');
 const nunjucks = require('nunjucks');
 const dateFilter = require('nunjucks-date-filter');
 const session = require('express-session');
-const redis = require('redis');
-const connectRedis = require('connect-redis');
 const pinoExpress = require('express-pino-logger');
 const uuid = require('uuid');
 const fileUpload = require('express-fileupload');
+const sessionConfig = require('./lib/session');
 require('express-async-errors');
 
 const config = require('./config');
@@ -19,14 +18,6 @@ const routes = require('./routes');
 
 const app = express();
 
-const RedisStore = connectRedis(session);
-
-const { sessionSecret } = config.server;
-
-if (!sessionSecret) {
-  throw new Error('Session secret must be set');
-}
-
 app.use(
   pinoExpress({
     logger,
@@ -34,19 +25,8 @@ app.use(
   })
 );
 
-const sessionConfig = {
-  store: new RedisStore({
-    client: redis.createClient(config.redis()),
-  }),
-  secret: sessionSecret,
-  resave: false,
-  saveUninitialized: true,
-  cookie: {},
-};
-
 if (config.server.useSecureSessionCookie) {
   app.set('trust proxy', 1); // trust first proxy
-  sessionConfig.cookie.secure = true; // serve secure cookies
 }
 
 app.use(compression());
@@ -55,7 +35,7 @@ app.use(lusca.xssProtection(true));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(session(sessionConfig));
+app.use(session(sessionConfig()));
 app.use('/public', express.static(path.join(__dirname, 'public')));
 app.use(
   '/assets',
